Move decompress methods to DecompressUtils
